chore: 将项目管理相关代码移动到 ProjectVM
- 优化撤回重做，不使用 ApplicationCommands 以和 TextBox 的撤回重做隔离
