+++ src/S7SvrSim/ViewModels/MainVM.cs
@@ -17,8 +17,6 @@ namespace S7Svr.Simulator.ViewModels
     public class MainVM : ReactiveObject
     {
         private readonly IS7ServerService _server;
-        private readonly IProjectFactory projectFactory;
-        private IProject currentProject;
 
         public ConfigSnap7ServerVM ConfigVM { get; }
         public MsgLoggerVM LoggerVM { get; }
@@ -26,24 +24,11 @@ namespace S7Svr.Simulator.ViewModels
         public RunningSnap7ServerVM RunningVM { get; }
         public OperationVM OperationVM { get; }
         public SignalWatchVM SignalWatchVM { get; }
-
-        [Reactive]
-        public bool NeedSave { get; set; }
-
-        [Reactive]
-        public int UndoCount { get; set; }
-
-        [Reactive]
-        public int RedoCount { get; set; }
-
-        public string ProjectTitle => Path.GetFileName(currentProject.Path);
+        public ProjectVM ProjectVM { get; }
 
         public MainVM(IS7ServerService server, IProjectFactory projectFactory)
         {
             this._server = server;
-            this.projectFactory = projectFactory;
-
-            currentProject = projectFactory.GetOrCreateProject(null);
 
             this.LoggerVM = Locator.Current.GetRequiredService<MsgLoggerVM>();
             this.ConfigPyEngineVM = Locator.Current.GetRequiredService<ConfigPyEngineVM>();
@@ -51,17 +36,12 @@ namespace S7Svr.Simulator.ViewModels
             this.OperationVM = Locator.Current.GetRequiredService<OperationVM>();
             this.ConfigVM = Locator.Current.GetRequiredService<ConfigSnap7ServerVM>();
             this.SignalWatchVM = Locator.Current.GetRequiredService<SignalWatchVM>();
+            this.ProjectVM = Locator.Current.GetRequiredService<ProjectVM>();
 
             var watchRunningStatus = this.WhenAnyValue(vm => vm.RunningVM.RunningStatus);
             this.CmdStartServer = ReactiveCommand.CreateFromTask(CmdStartServer_Impl, watchRunningStatus.Select(r => !r));
             this.CmdStopServer = ReactiveCommand.CreateFromTask(CmdStopServer_Impl, watchRunningStatus);
             this.CmdRestartServer = ReactiveCommand.CreateFromTask(CmdRestartServer_Impl, watchRunningStatus);
-
-            UndoRedoManager.UndoRedoChanged += () =>
-            {
-                UndoCount = UndoRedoManager.UndoCount;
-                RedoCount = UndoRedoManager.RedoCount;
-            };
         }
 
         #region 启停服务器
@@ -116,121 +96,5 @@ namespace S7Svr.Simulator.ViewModels
             await CmdStartServer_Impl();
         }
         #endregion
-
-        private void CallbackNeedSave()
-        {
-            NeedSave = true;
-            NeedSave = false;
-        }
-
-        public MessageBoxResult? NotifyIfSave()
-        {
-            if (NeedSave)
-            {
-                var result = MessageBox.Show("当前项目未保存，是否保存？", "未保存项目", MessageBoxButton.YesNoCancel);
-                if (result == MessageBoxResult.Yes)
-                {
-                    SaveProject();
-                }
-                return result;
-            }
-            return null;
-        }
-
-        public void OpenProjectFolder()
-        {
-            Process.Start("explorer.exe", $"/select,{currentProject.Path}");
-        }
-
-        public void NewProject()
-        {
-            if (NotifyIfSave() == MessageBoxResult.Cancel)
-            {
-                return;
-            }
-
-            SaveFileDialog saveFileDialog = new SaveFileDialog()
-            {
-                Title = "选择新项目保存路径",
-                Filter = $"S7模拟项目|*{ProjectConst.FILE_EXTENSION}",
-                FileName = Path.GetFileName(currentProject.Path),
-                RestoreDirectory = true,
-            };
-
-            if (saveFileDialog.ShowDialog() != true)
-            {
-                return;
-            }
-
-            currentProject = projectFactory.CreateProject(saveFileDialog.FileName);
-            UndoRedoManager.Reset();
-
-            CallbackNeedSave();
-        }
-
-        public void LoadProject()
-        {
-            if (NotifyIfSave() == MessageBoxResult.Cancel)
-            {
-                return;
-            }
-
-            OpenFileDialog openFileDialog = new OpenFileDialog()
-            {
-                Title = "选择项目文件",
-                Filter = $"S7模拟项目|*{ProjectConst.FILE_EXTENSION}",
-                Multiselect = false,
-                RestoreDirectory = true,
-            };
-
-            if (openFileDialog.ShowDialog() != true)
-            {
-                return;
-            }
-
-            IProject project;
-
-            try
-            {
-                project = projectFactory.GetProject(openFileDialog.FileName);
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show($"{ex}", "打开项目文件失败", MessageBoxButton.OK, MessageBoxImage.Error);
-                return;
-            }
-
-            currentProject = project;
-            UndoRedoManager.Reset();
-
-            CallbackNeedSave();
-        }
-
-        public void SaveProject()
-        {
-            currentProject.Save();
-            NeedSave = false;
-        }
-
-        public void SaveProjectAs()
-        {
-            SaveFileDialog saveFileDialog = new SaveFileDialog()
-            {
-                Title = "选择新项目保存路径",
-                Filter = $"S7模拟项目|*{ProjectConst.FILE_EXTENSION}",
-                FileName = Path.GetFileName(currentProject.Path),
-                RestoreDirectory = true,
-            };
-
-            if (saveFileDialog.ShowDialog() != true)
-            {
-                return;
-            }
-
-            currentProject.SaveAs(saveFileDialog.FileName);
-            currentProject = projectFactory.GetProject(saveFileDialog.FileName);
-
-            CallbackNeedSave();
-        }
     }
 }
